+++ quint/src/runtime/impl/compilerImpl.ts
@@ -873,18 +873,27 @@ export class CompilerVisitor implements IRVisitor {
       this.addCompileError(app.id, `Called unknown operator ${app.opcode}`)
       this.compStack.push(fail)
     } else {
+      const nparams = callable.registers.length
+      // nparams === nargs, unless a tuple is passed to an n-ary operator.
+      // The type checker should have checked the types before.
+      const nargs = (app.args.length === 1 && nparams > 1) ? 1 : nparams
+
       const nactual = this.compStack.length
-      const nexpected = callable.registers.length
-      if (nactual < nexpected) {
+      if (nactual < nargs) {
         this.addCompileError(app.id,
-          `Expected ${nexpected} arguments for ${app.opcode}, found: ${nactual}`)
+          `Expected ${nargs} arguments for ${app.opcode}, found: ${nactual}`)
         this.compStack.push(fail)
       } else {
         this.applyFun(app.id,
-          callable.registers.length,
+          nargs,
           (...args: RuntimeValue[]) => {
-            for (let i = 0; i < args.length; i++) {
-              callable.registers[i].registerValue = just(args[i])
+            let actualArgs = args
+            if (nparams > nargs && args.length === 1) {
+              // unpack the tuple
+              actualArgs = [...args[0]]
+            }
+            for (let i = 0; i < actualArgs.length; i++) {
+              callable.registers[i].registerValue = just(actualArgs[i])
             }
             return callable.eval() as Maybe<RuntimeValue>
           }
@@ -988,17 +997,23 @@ export class CompilerVisitor implements IRVisitor {
     }
     // lambda translated to Callable
     const callable = this.compStack.pop() as Callable
-    // this method supports only 1-argument callables
-    if (callable.registers.length !== 1) {
-      const nargs = callable.registers.length
-      this.addCompileError(sourceId, `Expected 1 argument, found ${nargs}`)
-      return
-    }
+    const nargs = callable.registers.length
     // apply the lambda to a single element of the set
     const evaluateElem = function(elem: RuntimeValue):
       Maybe<[RuntimeValue, RuntimeValue]> {
-      // store the set element in the register
-      callable.registers[0].registerValue = just(elem)
+      if (nargs === 1) {
+        // store the set element in the register
+        callable.registers[0].registerValue = just(elem)
+      } else {
+        // unpack a tuple and store its elements in the registers
+        const tupleElems = [...elem]
+        if (tupleElems.length !== nargs) {
+          return none()
+        }
+        for (let i = 0; i < nargs; i++) {
+          callable.registers[i].registerValue = just(tupleElems[i])
+        }
+      }
       // evaluate the predicate using the register
       // (cast the result to RuntimeValue, as we use runtime values)
       const result = callable.eval().map(e => e as RuntimeValue)