+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - short-circuiting for `implies` (#717)
+- improve the summary output of `run` (#719)
+- support for tuple unpacking in the simulator (#720)
 
 ### Changed
 ### Deprecated